Move Font,Music,Renderer,Texture,Timer out
