+++ MiniEngine.h
@@ -9,15 +9,7 @@
 
 namespace MiniEngine
 {
-	class Point
-	{
-	public:
-		int x, y;
-		Point(int X, int Y);
-		Point();
-		SDL_Point toSDLPoint() const;
-		bool inRect(const Rect& rect) const;
-	};
+
 
 	class NonCopyable
 	{
@@ -60,40 +52,7 @@ namespace MiniEngine
 
 	enum class BlendMode { None,Blend,Add,Mod };
 
-	class Texture
-	{
-	public:
-	    Texture();
-		~Texture() = default;
-		Rect getSize();
-		int getw() const;
-		int geth() const;
-		bool isReady() const;
-		int setBlendMode(BlendMode mode);
-		BlendMode getBlendMode() const;
-		/// Alpha:  0: Transparent 255: opaque
-		int setAlphaMode(int alpha);
-		int getAlphaMode() const;
 
-		ColorMode getColorMode() const;
-		int setColorMode(ColorMode mode);
-		RGBA getRGBA() const;
-		void setRGBA(const RGBA& pack);
-
-		void release();
-	protected:
-		/// updateInfo() must be called after Texture is changed.
-		void updateInfo();
-	private:
-		std::shared_ptr<SDL_Texture> _text;
-		void _set(SDL_Texture*);
-		/// Just used for "SDL_GetRenderTarget"
-		void _set_no_delete(SDL_Texture*);
-		void _clear();
-		SDL_Texture* _get() const;
-		Rect rect;
-		friend class Renderer;
-	};
 
 	enum class SystemCursorType
 	{
@@ -239,249 +198,11 @@ namespace MiniEngine
 
 	enum class FlipMode { None, Horizontal, Vertical };
 
-	class Renderer
-	{
-	public:
-	    Renderer() = default;
-	    Renderer(Window& wnd,std::initializer_list<RendererType> RendererFlags = { RendererType::Accelerated,RendererType::TargetTexture }) throw (ErrorViewer);
-	    ~Renderer() = default;
 
-	    /// If Renderer is current not ready, setRenderer will create Renderer.
-	    /// Otherwise, setRenderer will fail.
-	    int createRenderer(Window& wnd,RendererType Type)
-		{
-		    int flagcalc=0;
-            return _createRenderer(wnd,flagcalc,Type);
-		}
-
-	    template<typename... Args>
-		int createRenderer(Window& wnd,RendererType Type,Args&&... args)
-		{
-            int flagcalc=0;
-            return _createRenderer(wnd,flagcalc,Type,std::forward<RendererType>(args...));
-		}
-
-		int createRenderer(Window& wnd,std::initializer_list<RendererType>);
-
-		int setColor(const RGBA& pack);
-		RGBA getColor() const;
-		int setBlendMode(BlendMode mode);
-		BlendMode getBlendMode() const;
-
-		int setTarget(Texture& t);
-		int setTarget();
-		Texture getTarget();
-
-		int fillRect(const Rect& rect);
-		int drawRect(const Rect& rect);
-		int drawPoint(const Point& p);
-		int drawLine(const Point& A,const Point& B);
-
-		/// Experimental
-        int fillRects(const SDL_Rect* pRectArray,int n);
-        int drawRects(const SDL_Rect* pRectArray,int n);
-        int drawPoints(const SDL_Point* pPointArray,int n);
-        int drawLines(const SDL_Point* pPointArray,int n);
-        /// Experimental
-        int fillRects(const std::vector<SDL_Rect>& rectvec);
-        int drawRects(const std::vector<SDL_Rect>& rectvec);
-        int drawPoints(const std::vector<SDL_Point>& pointvec);
-        int drawLines(const std::vector<SDL_Point>& pointvec);
-
-        int setScale(float scaleX,float scaleY);
-        std::tuple<float,float> getScale() const;
-
-        int setViewport(const Rect& viewport);
-        Rect getViewport() const;
-
-        int setLogicalSize(int w,int h);
-        Rect getLogicalSize() const;
-
-        int setClipRect(const Rect& cliprect);
-        Rect getClipRect() const;
-        bool isClipEnabled() const;
-
-        Rect getOutputSize() const;
-
-		int clear();
-		void update();
-
-		int copy(const Texture& t, const Rect& src, const Rect& dst);
-		int copyTo(const Texture& t, const Rect& dst);
-		int copyTo(const Texture& t, const Point& lupoint);
-		int copyFill(const Texture& t, const Rect& src);
-		int copyFullFill(const Texture& t);
-
-		/// Super copy without center point.
-		int copy(const Texture& t, const Rect& src, const Rect& dst,double angle,FlipMode mode);
-		int copyTo(const Texture& t, const Rect& dst,double angle,FlipMode mode);
-		int copyTo(const Texture& t, const Point& lupoint,double angle,FlipMode mode);
-		int copyFill(const Texture& t, const Rect& src,double angle,FlipMode mode);
-		int copyFullFill(const Texture& t,double angle,FlipMode mode);
-		/// Super copy with center point
-		int copy(const Texture& t, const Rect& src, const Rect& dst,const Point& centerPoint,double angle,FlipMode mode);
-		int copyTo(const Texture& t, const Rect& dst,const Point& centerPoint,double angle,FlipMode mode);
-		int copyTo(const Texture& t, const Point& lupoint,const Point& centerPoint,double angle,FlipMode mode);
-		int copyFill(const Texture& t, const Rect& src,const Point& centerPoint,double angle,FlipMode mode);
-		int copyFullFill(const Texture& t,const Point& centerPoint,double angle,FlipMode mode);
-
-		/// Reserved for compatibility
-		int supercopy(const Texture& t,
-                bool srcfull,const Rect& src,bool dstfull,const Rect& dst,
-                double angle,
-                bool haspoint,const Point& center,FlipMode mode);
-
-		Texture render(const Surface& surf) const throw (ErrorViewer);
-		Texture loadTexture(const std::string& FileName) const throw(ErrorViewer);
-		Texture loadTextureRW(const RWOP& rwop) const throw(ErrorViewer);
-		Texture createTexture(int Width, int Height) const throw(ErrorViewer);
-
-		bool isRenderTargetSupported() const;
-		bool isReady() const;
-
-		void release();
-
-		/// Experimental
-		static int GetDriversNum();
-    protected:
-        template<typename... Args>
-        int _createRenderer(Window& wnd,int& refcalc,RendererType Type,Args&&... args)
-        {
-            refcalc|=_rendertype_caster(Type);
-            return _createRenderer(wnd,refcalc,args...);
-        }
-
-        // template<>
-        int _createRenderer(Window& wnd,int& refcalc,RendererType Type)
-        {
-            refcalc|=_rendertype_caster(Type);
-            return _createRenderer_Real(wnd,refcalc);
-        }
-	private:
-	    std::shared_ptr<SDL_Renderer> _rnd;
-
-        int _createRenderer_Real(Window& wnd,Uint32 flags);
-		Uint32 _rendertype_caster(RendererType);
-
-		void _set(SDL_Renderer*);
-		void _clear();
-		SDL_Renderer* _get() const;
-	};
 
 	enum class FontStyle { Normal, Bold, Italic, UnderLine, StrikeThrough };
 	enum class FontHint { Normal, Light, Mono, None , Error };
 
-	class Font
-	{
-	public:
-		Font() = default;
-		Font(std::string FontFileName, size_t size) throw(ErrorViewer);
-		int use(std::string FontFileName, size_t size);
-		bool isReady() const;
-
-		bool isNormal() const;
-        bool isBold() const;
-        bool isItalic() const;
-        bool isUnderLine() const;
-        bool isStrikeThrough() const;
-
-        void setNormal();
-        void setBold(bool);
-        void setItalic(bool);
-        void setUnderLine(bool);
-        void setStrikeThrough(bool);
-
-		template<typename... Args>
-		void setFontStyle(FontStyle style,Args&&... args)
-		{
-		    int fontcalc=0;
-		    _setFontStyle(fontcalc,style,args...);
-		}
-
-		void setFontStyle(FontStyle style)
-		{
-            int fontcalc=0;
-            _setFontStyle(fontcalc,style);
-		}
-
-		std::vector<FontStyle> getFontStyles() const;
-
-		int getFontHeight() const;
-		int getFontAscent() const;
-		int getFontDescent() const;
-		int getFontLineSkip() const;
-
-		bool isFontKerning() const;
-		void setFontKerning(bool enableKerning);
-
-		long getFontFaceNum() const;
-        int getFontFaceIsFixedWidth() const;
-        std::string getFontFaceFamilyName() const;
-        std::string getFontFaceStyleName() const;
-
-		FontHint getFontHint() const;
-		void setFontHint(FontHint hint);
-
-
-        Rect sizeText(const std::string& Text) const throw (ErrorViewer);
-        Rect sizeUTF8(const std::string& Text) const throw (ErrorViewer);
-        Rect sizeUnicode(const uint16_t* Text) const throw (ErrorViewer);
-
-        /// Surface Rendering Functions.
-		Surface renderText(const std::string& Text, const RGBA& fg) const;
-		Surface renderTextWrapped(const std::string& Text, const RGBA& fg, size_t WrapLength) const;
-		Surface renderTextShaded(const std::string& Text, const RGBA& fg, const RGBA& bg) const;
-		Surface renderTextSolid(const std::string& Text, const RGBA& fg) const;
-
-		Surface renderUTF8(const std::string& Text, const RGBA& fg) const;
-		Surface renderUTF8Wrapped(const std::string& Text, const RGBA& fg, size_t WrapLength) const;
-		Surface renderUTF8Shaded(const std::string& Text, const RGBA& fg, const RGBA& bg) const;
-		Surface renderUTF8Solid(const std::string& Text, const RGBA& fg) const;
-
-		Surface renderUnicode(const uint16_t* Text,const RGBA& fg) const;
-		Surface renderUnicodeWrapped(const uint16_t* Text,const RGBA& fg,size_t WrapLength) const;
-		Surface renderUnicodeShaded(const uint16_t* Text,const RGBA& fg,const RGBA& bg) const;
-		Surface renderUnicodeSolid(const uint16_t* Text,const RGBA& fg) const;
-
-		/// Texture Rendering Functions.
-		Texture renderText(const Renderer& rnd, const std::string& Text, const RGBA& fg) const;
-		Texture renderTextWrapped(const Renderer& rnd, const std::string& Text, const RGBA& fg, size_t WrapLength) const;
-		Texture renderTextShaded(const Renderer& rnd, const std::string& Text, const RGBA& fg, const RGBA& bg) const;
-		Texture renderTextSolid(const Renderer& rnd, const std::string& Text, const RGBA& fg) const;
-
-		Texture renderUTF8(const Renderer& rnd, const std::string& Text, const RGBA& fg) const;
-		Texture renderUTF8Wrapped(const Renderer& rnd, const std::string& Text, const RGBA& fg, size_t WrapLength) const;
-		Texture renderUTF8Shaded(const Renderer& rnd, const std::string& Text, const RGBA& fg, const RGBA& bg) const;
-		Texture renderUTF8Solid(const Renderer& rnd, const std::string& Text, const RGBA& fg) const;
-
-		Texture renderUnicode(const Renderer& rnd,const uint16_t* Text,const RGBA& fg) const;
-		Texture renderUnicodeWrapped(const Renderer& rnd,const uint16_t* Text,const RGBA& fg,size_t WrapLength) const;
-		Texture renderUnicodeShaded(const Renderer& rnd,const uint16_t* Text,const RGBA& fg,const RGBA& bg) const;
-		Texture renderUnicodeSolid(const Renderer& rnd,const uint16_t* Text,const RGBA& fg) const;
-
-		void release();
-    protected:
-		template<typename... Args>
-        void _setFontStyle(int& fontcalc,FontStyle style,Args&&... args)
-        {
-            fontcalc|=_style_caster(style);
-            _setFontStyle(fontcalc,args...);
-        }
-
-        void _setFontStyle(int& fontcalc,FontStyle style)
-        {
-            fontcalc|=_style_caster(style);
-            _real_setFontStyle(fontcalc);
-        }
-	private:
-	    void _real_setFontStyle(int);
-	    int _style_caster(FontStyle);
-
-		std::shared_ptr<TTF_Font> _font;
-		void _set(TTF_Font*);
-		void _clear();
-		TTF_Font* _get() const;
-	};
 
 	enum class Platform { Unknown,Windows,MacOS,Linux,iOS,Android };
 	enum class PowerState { Unknown,OnBattery,NoBattery,Charging,Charged };
@@ -578,155 +299,8 @@ namespace MiniEngine
 		};
 	};
 
-	Uint32 _global_timer_executor(Uint32 interval,void* param);
 
-	class Timer
-	{
-    public:
-        Timer();
 
-        /// void func(Uint32,...)
-        template<typename VoidCallable,typename... Args>
-        Timer(Uint32 interval,VoidCallable&& vcallable,Args&&... args) : Timer()
-        {
-            auto realCall=[&,vcallable](Uint32 ims)->Uint32{vcallable(ims,args...);return ims;};
-            auto pfunc=new std::function<Uint32(Uint32 interval)>(realCall);
-            _real_timer_call(_global_timer_executor,interval,pfunc);
-        }
-
-        /// Uint32 func(Uint32,...)
-        template<typename Callable,typename... Args>
-        Timer(Callable&& callable,Uint32 interval,Args&&... args) : Timer()
-        {
-            auto realCall=[&,callable](Uint32 ims)->Uint32{return callable(ims,args...);};
-            auto pfunc=new std::function<Uint32(Uint32 interval)>(realCall);
-            _real_timer_call(_global_timer_executor,interval,pfunc);
-        }
-
-        /// Restore For Capability
-        Timer(SDL_TimerCallback callback,Uint32 interval,void* param);
-
-        int enable();
-        int disable();
-        bool isenable() const;
-        void detach();
-        ~Timer();
-
-        static void _delete_delegator(std::function<Uint32(Uint32)>* Delegator);
-    private:
-
-        void _real_timer_call(SDL_TimerCallback callback,Uint32 interval,void* param);
-
-        SDL_TimerCallback _callback;
-        Uint32 _interval;
-        void* _param;
-        SDL_TimerID id;
-        bool _enabled;
-        bool _detached;
-        /// Reserved Variable For Template variable Parameter
-        bool _delete_on_disable;
-	};
-
-	class AudioPlayer
-	{
-	public:
-		AudioPlayer();
-		~AudioPlayer();
-	private:
-		class _Audio
-		{
-		public:
-			_Audio();
-			~_Audio();
-		};
-
-		static _Audio* _sysAudio;
-		static int _sysAudioCounter;
-	};
-
-	/// Forward Declaration
-	class Music
-	{
-	public:
-
-	protected:
-		Music() = default;
-	private:
-		std::shared_ptr<Mix_Music> _music;
-        void _set(Mix_Music*);
-        void _clear();
-        Mix_Music* _get();
-		friend class MusicPlayer;
-	};
-
-	class MusicPlayer : public AudioPlayer
-	{
-	public:
-	    static int GetDecoderNum();
-		static std::string GetDecoderName(int index);
-
-		Music loadMusic(std::string Filename) throw (ErrorViewer);
-
-		int play(Music music, int loops);
-		void pause();
-		void resume();
-		void rewind();
-		int stop();
-		int fadeIn(int loops, int ms);
-		int fadeOut(int ms);
-
-		bool isPlaying();
-		bool isPaused();
-		int isFading();
-
-		/// Experimental
-		static int SetMusicPosition(double position);
-
-	private:
-		Music m;
-	};
-
-	class Sound
-	{
-	public:
-	protected:
-		Sound() = default;
-	private:
-		std::shared_ptr<Mix_Chunk> _sound;
-		void _set(Mix_Chunk*);
-		void _clear();
-		Mix_Chunk* _get();
-		friend class SoundPlayer;
-	};
-
-	typedef int ChannelID;
-
-	class SoundPlayer : public AudioPlayer
-	{
-	public:
-	    static int GetDecoderNum();
-		static std::string GetDecoderName(int index);
-
-		SoundPlayer(int Channels = 16);
-		Sound loadSound(std::string Filename) throw (ErrorViewer);
-		ChannelID playSound(Sound sound, int loops) throw (ErrorViewer);
-		ChannelID fadein(Sound sound, int loops, int ms) throw (ErrorViewer);
-		int fadeout(ChannelID id, int ms);
-		void pause(ChannelID id);
-		void resume(ChannelID id);
-		int stop(ChannelID id);
-
-		/// Experimental
-		int setPanning(ChannelID id,uint8_t left,uint8_t right);
-		int setPosition(ChannelID id,int16_t angle,uint8_t distance);
-		int setDistance(ChannelID id,uint8_t distance);
-		int setReverseStereo(ChannelID id,int flip);
-
-		/// Experimental: Direct Add/Remove Effect
-		int addEffect(ChannelID id,Mix_EffectFunc_t f, Mix_EffectDone_t d, void *arg);
-		int removeEffect(ChannelID id,Mix_EffectFunc_t f);
-		int removeAllEffect(ChannelID id);
-	};
 
 	class StringEngine
 	{
